feat: select worry view model
